Fix packing algorithm
Take a different approach to the problem. I now realize this is basically the change making problem but with the twist of overshooting to meat the target amount. I opted for the dynamic programming approach as the greedy approach does not meet the requirements of the solution (minimum num of items, and minimum number of packs)